+++ js/userInfo.js
@@ -1,0 +1,4 @@
+export const user = {
+  email: "test@codeit.com",
+  password: "codeit101"
+};
